Changes how additional web preferences are generated.
The app installer now writes WebPreferences defined in the appconfig to the topic text instead of writing them as metadata.

Ticket: none

Release-Note(change,admin){
All WebPreferences which are set during installation with the appmanager are now set in the WebPreferences topic text.
Previously they were saved in the topic metadata and only accessible via the WebPreferences site config.
}